move editor eligiblity functions from users.models to users.helpers.
